Add a hack to insert required whitespace in generated AsciiDoc
A better solution would be to have an actual tree traversal, with the
rules about how the various elements must be combined, but for now just
add spaces and new lines before the markup elements requiring them.

Note that this doesn't work for the header markup, as the "$doc" is
still empty when they're processed at the top level, so we still need to
prepend them with a new line.